Load data files
If another user wants to load the data files, the path directing to the data files needs to be changed to the user's directory.
Moreover, one has to make sure the data file has the same structure as described in the comments. That is: 
The data file called 'score' has to contain the 'SID' in the first column and the Activity Score in the second column.
The data file called 'descriptors' has to contain the 'SID' in the first column and the descriptors in the remaining columns.